requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
async-timeout==4.0.3
asyncio==4.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dataclasses==0.6
dataclasses-json==0.6.7
ddgs==9.5.4
distro==1.9.0
environs==9.5.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
faiss-cpu==1.12.0
filelock==3.19.1
Flask==2.3.3
flask-cors==6.0.1
Flask-SocketIO==5.3.6
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langsmith==0.4.14
lxml==6.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
milvus-lite==2.5.1
minio==7.2.16
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
narwhals==2.1.2
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.99.9
openpyxl==3.1.5
orjson==3.11.2
packaging==23.2
pandas==2.3.1
pillow==11.3.0
pluggy==1.6.0
primp==0.15.0
propcache==0.3.2
protobuf==6.32.0
psycopg2-binary==2.9.10
pyarrow==21.0.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pymilvus==2.6.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.7.1
python-socketio==5.9.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.27.0
safetensors==0.6.2
schedule==1.2.2
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==5.1.0
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
streamlit==1.48.1
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
toml==0.10.2
tomli==2.2.1
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.55.2
triton==3.4.0
typing==3.7.4.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
watchdog==6.0.0
websocket-client==1.8.0
Werkzeug==2.3.7
wsproto==1.2.0
yarl==1.20.1
zstandard==0.23.0
